Multi table serializer
